impl part2 add comment model/repository/crud service

+++ hw06-spring-jpa/src/main/java/ru/otus/hw06/impl/repositories/GenreRepositoryJpaImpl.java
@@ -20,7 +20,7 @@ public class GenreRepositoryJpaImpl implements GenreRepositoryJpa {
 
   @Override
   @Transactional(readOnly = true)
-  public int count() {
+  public long count() {
     String sql = "select count(g) from Genre g";
     return em.createQuery(sql, int.class).getSingleResult();
   }
